@@ -41,10 +41,10 @@
     "lottie-react": "^2.1.0",
     "material-ui-popup-state": "^1.9.3",
     "qr-code-styling": "^1.6.0-rc.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.1",
     "react-data-grid": "^7.0.0-canary.49",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^12.0.4",
     "react-grid-layout": "^1.3.0",
     "react-i18next": "^11.8.12",
@@ -159,9 +159,9 @@
     "@types/lodash": "^4.14.168",
     "@types/node": "^12.0.0",
     "@types/qrcode-svg": "^1.1.1",
-    "@types/react": "^17.0.3",
+    "@types/react": "^18.2.19",
     "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.2.7",
     "@types/react-grid-layout": "^1.1.2",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-swipeable-views": "^0.13.0",
