@@ -1,5 +1,4 @@
 import { Provider } from 'react-redux'
-import ReactDOM from 'react-dom'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import { firebaseProps, persistor, store, TimeoutCheckProvider } from '@loopring-web/core'
@@ -15,6 +14,7 @@ import { PersistGate } from 'redux-persist/integration/react'
 import { provider, ProviderComposer, useSettings } from '@loopring-web/component-lib'
 import React, { Provider as TProvider } from 'react'
 import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
+import { createRoot } from 'react-dom/client'
 
 if (process.env.REACT_APP_VER) {
   console.log('VER:', process.env.REACT_APP_VER)
@@ -41,7 +41,8 @@ const ProviderThen = React.memo(({ children }: { children: JSX.Element }) => {
   return <ProviderComposer providers={providers}>{children}</ProviderComposer>
 })
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root') as HTMLElement)
+root.render(
   <ProviderApp>
     <ReactReduxFirebaseProvider {...firebaseProps}>
       <ProviderThen>
@@ -49,7 +50,6 @@ ReactDOM.render(
       </ProviderThen>
     </ReactReduxFirebaseProvider>
   </ProviderApp>,
-  document.getElementById('root'),
 )
 
 // If you want to start measuring performance in your app, pass a function
