@@ -1,5 +1,4 @@
 import { Provider } from 'react-redux'
-import ReactDOM from 'react-dom'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import { firebaseProps, persistor, store, TimeoutCheckProvider } from '@loopring-web/core'
@@ -15,6 +14,7 @@ import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
 import { LocalizationProvider } from '@mui/lab'
 import MomentUtils from '@mui/lab/AdapterMoment'
 import { I18nextProvider } from 'react-i18next'
+import { createRoot } from 'react-dom/client'
 
 if (process.env.REACT_APP_VER) {
   console.log('VER:', process.env.REACT_APP_VER)
@@ -40,7 +40,8 @@ const ProviderThen = React.memo(({ children }: { children: JSX.Element }) => {
   return <ProviderComposer providers={providers}>{children}</ProviderComposer>
 })
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root') as HTMLElement)
+root.render(
   <ProviderApp>
     <ReactReduxFirebaseProvider {...firebaseProps}>
       <ProviderThen>
@@ -48,7 +49,6 @@ ReactDOM.render(
       </ProviderThen>
     </ReactReduxFirebaseProvider>
   </ProviderApp>,
-  document.getElementById('root'),
 )
 
 // If you want to start measuring performance in your app, pass a function
